server/requirements.txt
asn1crypto=0.24.0
cffi=1.11.5
click=6.7
Flask=0.12.2
Flask-SQLAlchemy=2.3.2
idna=2.6
itsdangerous=0.24
Jinja2=2.10
MarkupSafe=1.0
PyJWT=1.6.1
PyMySQL=0.8.0
setuptools=39.0.1
six=1.11.0
SQLAlchemy=1.2.6
Werkzeug=0.14.1
wheel=0.31.0
